CyberSecMonitor/requirements.txt
Django>=5.2
djangorestframework>=3.15
channels>=4.1
django-filter>=24.3
python-dotenv>=1.0
aiohttp>=3.9
pandas>=2.2
numpy>=2.0
matplotlib>=3.8
daphne>=4.1
reportlab>=4.0
